rename Step and Steps to lowercase
