+++ src/components/legends/BerthLegend/index.tsx
@@ -7,7 +7,7 @@ import Icon, { IconNames } from '../../common/Icon';
 import AutoSave from '../../forms/AutoSave';
 import Form from '../../forms/Form';
 import ApplicationSelector from '../../forms/sections/applicationSelector/ApplicationSelector';
-import Steps from '../../steps';
+import Steps from '../../steps/Steps';
 
 import {
   BerthsServices,
